+++ modules/sql-engine/src/main/java/org/apache/ignite/internal/sql/engine/prepare/DdlPlan.java
@@ -20,8 +20,8 @@ package org.apache.ignite.internal.sql.engine.prepare;
 import java.util.Collections;
 import java.util.List;
 import org.apache.ignite.internal.catalog.CatalogCommand;
-import org.apache.ignite.internal.sql.api.ColumnMetadataImpl;
-import org.apache.ignite.internal.sql.api.ResultSetMetadataImpl;
+import org.apache.ignite.internal.sql.ColumnMetadataImpl;
+import org.apache.ignite.internal.sql.ResultSetMetadataImpl;
 import org.apache.ignite.internal.sql.engine.SqlQueryType;
 import org.apache.ignite.sql.ColumnMetadata;
 import org.apache.ignite.sql.ColumnType;
